ADD: New authentication mechanism
Rather than having the user enter their login information into the
application, this change makes it so that the authentication data that
Idle Master needs is extracted directly from the Steam application.
This change will also make it easier for users with multiple accounts,
as they'll be able to in as that account, launch Steam, and not need to
reconfigure Idle Master's authentication data.

This method currently doesn't work if Steam Family View is enabled, but
I'll try to make sure that's working before this version goes live (if
possible).